requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cloudinary==1.42.1
dj-database-url==2.3.0
Django==5.0.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
